requirements.txt
tensorflow==2.14.0
streamlit==1.31.0
numpy==1.24.3
Pillow==9.5.0
matplotlib==3.7.1
scikit-learn==1.2.2
